+++ backend/src/main/java/zipgo/common/config/WebConfig.java
@@ -8,7 +8,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import zipgo.auth.presentation.AuthInterceptor;
 import zipgo.auth.presentation.JwtArgumentResolver;
-import zipgo.auth.util.JwtProvider;
+import zipgo.auth.support.JwtProvider;
 
 import java.util.List;
